🔥 [yw_clients] Get rid of abstract `RequestExecutor`
