+++ lib/yw_clients/yw_clients/http/__init__.py
@@ -4,4 +4,3 @@ This module gathers HTTP clients related to youwol.
 
 # relative
 from .aiohttp_utils import *
-from .request_executor import *
